[AIRFLOW-5197] Move GCP Datastore to core
This commit moves GCP Datastore from contrib to core.
For more information check AIP-21.
